Baptista Cipher/requirements.txt
scikit-image
pillow
scipy